requirements.txt
asgiref==3.8.1
Django==5.2
django-etc==1.4.0
django-hitcount==1.3.5
django-modeltranslation==0.19.16
django-widget-tweaks==1.5.0
pillow==11.2.1
python-decouple==3.8
sqlparse==0.5.3
